check query of the view  #2827
